requirements.txt
# Core dependencies
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.7
sentence-transformers==2.2.2
transformers==4.35.0
torch>=2.0.0

# Vector database
chromadb==0.4.15
faiss-cpu==1.7.4

# Document processing
pypdf2==3.0.1
python-docx==1.1.0
unstructured==0.12.2
markdown==3.5.1

# Web framework
streamlit==1.28.0
fastapi==0.104.1
uvicorn==0.24.0

# HTTP requests
requests==2.31.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
click==8.1.7

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1